runtime: a 16-byte object header (rc is uint32)
The header was 24 bytes -- rc(8) + flags(4) + 4 pad + drop(8). Making
rc a uint32 lets it and flags fill the 8 bytes before the drop pointer
with no padding, so the header is 16 bytes and a Node falls from a
48-byte slab class to 32: binary-trees 0.43x -> 0.40x, every other
bench flat. Four billion references to one object is not a real number
-- each is a pointer stored somewhere, 32 GB of them to reach 2^32 --
and CBMC re-verifies the drop-at-exactly-zero contract at the narrower
width.

The drop-pointer half of the diet is deliberately NOT taken. Folding
it into a type index would save the last 8 bytes but add a
drop_table[type](h) lookup on the release path, which is 26.7% of
binary-trees after the slab -- the same indirection the
drop-devirtualisation experiment measured and rejected this morning.
The pointer stays a pointer until a profile says the size, not the
lookup, is the wall.

Two things surfaced doing it. The models were symbolically executing
the SLAB (my earlier __CPROVER guard is not a macro CBMC defines), and
its chunk-carving loop blew the unwind bound; the fix is an explicit
-DNEON_CBMC from runtime/models/CMakeLists.txt that selects the
malloc path, so the models verify the refcount CONTRACT the slab
preserves rather than the allocator internals. And the header change
is otherwise invisible: everything goes through sizeof(neon_header)
and field access, the oracle is inert, 293 corpus programs pass under
ASan/UBSan/LSan, and eight refcount/list/map/resource/str CBMC models
pass at the new width.

+++ runtime/include/neon/core.h
@@ -19,8 +19,18 @@
 
 // Every heap object begins with this. `drop` frees *this* object (releasing its own
 // counted fields first); a NULL owner / an immortal flag makes retain/release no-ops.
+//
+// `rc` is `uint32_t`, not `uint64_t`, and it is load-bearing for size: with `flags` it
+// fills the eight bytes before the (8-byte-aligned) `drop` pointer exactly, so the header
+// is 16 bytes rather than 24 — a third off every heap object, and a Node drops from a
+// 48-byte slab class to 32. Four billion references to ONE object is not a real number:
+// each reference is a pointer stored somewhere, so reaching 2^32 would take 32 GB of
+// pointers to a single node. The `drop` pointer is kept a real pointer, deliberately —
+// folding it into a type index would trade eight bytes for a table lookup on the release
+// path, which is a quarter of binary-trees' run, exactly the indirection the drop-
+// devirtualisation experiment showed does not pay.
 typedef struct neon_header {
-    uint64_t rc;
+    uint32_t rc;
     uint32_t flags;
     void (*drop)(void*);
 } neon_header;
